// give-away/src/api/models/PointPackageListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Give Away API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PointPackageStatus } from './PointPackageStatus';
import {
    PointPackageStatusFromJSON,
    PointPackageStatusFromJSONTyped,
    PointPackageStatusToJSON,
} from './PointPackageStatus';

/**
 * 
 * @export
 * @interface PointPackageListResponse
 */
export interface PointPackageListResponse {
    /**
     * 
     * @type {string}
     * @memberof PointPackageListResponse
     */
    pointPackageId?: string;
    /**
     * 
     * @type {number}
     * @memberof PointPackageListResponse
     */
    points?: number;
    /**
     * 
     * @type {number}
     * @memberof PointPackageListResponse
     */
    price?: number;
    /**
     * 
     * @type {PointPackageStatus}
     * @memberof PointPackageListResponse
     */
    status?: PointPackageStatus;
}

/**
 * Check if a given object implements the PointPackageListResponse interface.
 */
export function instanceOfPointPackageListResponse(value: object): value is PointPackageListResponse {
    return true;
}

export function PointPackageListResponseFromJSON(json: any): PointPackageListResponse {
    return PointPackageListResponseFromJSONTyped(json, false);
}

export function PointPackageListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PointPackageListResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'pointPackageId': json['pointPackageId'] == null ? undefined : json['pointPackageId'],
        'points': json['points'] == null ? undefined : json['points'],
        'price': json['price'] == null ? undefined : json['price'],
        'status': json['status'] == null ? undefined : PointPackageStatusFromJSON(json['status']),
    };
}

export function PointPackageListResponseToJSON(value?: PointPackageListResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'pointPackageId': value['pointPackageId'],
        'points': value['points'],
        'price': value['price'],
        'status': PointPackageStatusToJSON(value['status']),
    };
}